task 1

+++ file_io/1-create_file.c
@@ -29,7 +29,7 @@ size_t _strlen(char *str)
 int create_file(const char *filename, char *text_content)
 {
 	int fd = 0;
-	size_t num_letters = 0;
+	int num_letters = 0;
 
 	if (filename == NULL)
 		return (-1);
